+++ src/index.js
@@ -1,2 +1,3 @@
 
 export { default as PrettyInputText } from './components/PrettyInputText';
+export { default as PrettyInputTextArea } from './components/PrettyInputTextArea';