+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 
+## v1.0.9
+
+### Added
+- New prop `labelText` added to set the text of the label instead of `inputValue`.
+- New prop `isEnabled` added to control whether the component is enabled or not.
+- New prop `onChange` added to pass a callback function invoked on input change event.
+- New prop `onKeyPress` added to pass a callback function invoked on every key press.
+- New prop `labelColor` added to set the color of the component label.
+- New prop `backgroundColor` added to set the background color of the component.
+- New prop `width` added to set the width (in pixels) of the component.
+
+### Changed
+- Prop `inputValue` is not required now.
+
 ## v1.0.4
 
 ### Added